backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
anthropic
python-dotenv
PyJWT
cryptography
Werkzeug
eventlet

Flask-Cors
